+++ react-app/src/components/Formik/Formik.module.css
@@ -36,6 +36,7 @@ input[type="text"],
 input[type="url"],
 input[type="password"] {
     width: 90%;
+    font-size: 1rem;
     border-color: black;
 }
